[RFC] Extract hint functions to use new Stream
